requirements.txt
trinsic-sdk~=1.4.0
